omnidocs/tasks/math_expression_extraction/extractors/UniMERNet/unimernet/requirements.txt
# Copyright (c) OpenDataLab (https://github.com/opendatalab/UniMERNet)
# 
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.


torch>=2.2.2
torchvision>=0.17.2
transformers>=4.42.4
omegaconf>=2.3.0
timm>=0.9.16
opencv-python>=4.6.0
albumentations>=1.4.4
pillow>=10.0.0
numpy>=1.24.0
pyyaml>=6.0.0
tqdm>=4.64.0
iopath>=0.1.9
